fix: strange error with user password getter
